requirements.txt
async-generator==1.10
attrs==22.1.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.9.14
cffi==1.15.1
charset-normalizer==2.1.1
cloudscraper==1.2.64
h11==0.13.0
idna==3.4
lxml==4.9.1
outcome==1.2.0
pycparser==2.21
pyparsing==3.0.9
PySocks==1.7.1
python-dotenv==0.21.0
requests==2.28.1
requests-toolbelt==0.9.1
selenium==4.4.3
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
trio==0.21.0
trio-websocket==0.9.2
urllib3==1.26.12
wsproto==1.2.0
